relevé : la fenêtre posée à la main compile à nouveau
`clamp` ne rend un double que si le receveur et ses deux bornes en sont
statiquement ; la borne haute passait par `math.max`, et le résultat
retombait sur `num`, que `scale` refuse. Le build iOS s'arrêtait là.

+++ lib/domain/room/scanned_room.dart
@@ -332,7 +332,10 @@ class ScannedRoom {
     if (wall == null) return null;
     final width = math.min(size.width, wall.width);
     final half = width / 2;
-    final t = (p - wall.start).dot(wall.along).clamp(half, math.max(half, wall.width - half));
+    // `clamp` ne rend un double que si ses deux bornes en sont statiquement :
+    // `toDouble` le garantit, sinon `scale` reçoit un num et la compilation
+    // échoue.
+    final t = (p - wall.start).dot(wall.along).clamp(half, math.max(half, wall.width - half)).toDouble();
     return RoomSurface(
       kind: RoomSurfaceKind.window,
       center: wall.start + wall.along.scale(t),
